notebook-runner/requirements.txt
langchain
langchain-community
langchain-google-genai==2.1.5
faiss-cpu
pypdf
gtts
requests
beautifulsoup4
twilio
pydub
google-cloud-speech
google-cloud-texttospeech
python-dotenv
openai-whisper
numpy
PyPDF2